fix: mock GenerateSbomService in flaky SPDX unsupported component test
  - Add quarkus-junit5-mockito test dependency for @InjectMock support in
  @QuarkusTest
  - Create UploadSpdxMockedSyftRestTest with mocked GenerateSbomService to eliminate
  registry auth flakiness
  - Move testUpload_SpdxWithUnsupportedComponent_RecordsInSubmissionFailures from
  UploadSpdxRestTest to new class
  - Remove SPDX_WITH_UNSUPPORTED constant from UploadSpdxRestTest
