+++ src/test/java/com/redhat/ecosystemappeng/morpheus/rest/UploadSpdxMockedSyftRestTest.java
@@ -1,0 +1,89 @@
+package com.redhat.ecosystemappeng.morpheus.rest;
+
+import static org.hamcrest.Matchers.*;
+
+import java.io.File;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.redhat.ecosystemappeng.morpheus.model.ParsedCycloneDx;
+import com.redhat.ecosystemappeng.morpheus.service.GenerateSbomService;
+
+import io.quarkus.test.InjectMock;
+import io.quarkus.test.junit.QuarkusTest;
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+
+/**
+ * SPDX upload tests that mock {@link GenerateSbomService} to avoid real syft/registry calls.
+ */
+@QuarkusTest
+public class UploadSpdxMockedSyftRestTest {
+
+    private static final String SPDX_WITH_UNSUPPORTED = "src/test/resources/devservices/spdx-sboms/spdx-with-unsupported-component.json";
+    private static final String CYCLONEDX_FIXTURE = "src/test/resources/devservices/cyclonedx-sboms/nmstate-rhel8-operator.json";
+    private static final String TEST_VULN_ID = "CVE-2021-4238";
+
+    @InjectMock
+    GenerateSbomService generateSbomService;
+
+    @BeforeEach
+    void setUp() throws Exception {
+        RestApiTestFixture.configureRestAssuredIfExternal();
+
+        ObjectMapper mapper = new ObjectMapper();
+        JsonNode sbomJson = mapper.readTree(new File(CYCLONEDX_FIXTURE));
+        JsonNode component = sbomJson.get("metadata").get("component");
+
+        ParsedCycloneDx mockSbom = new ParsedCycloneDx(
+            sbomJson,
+            component.get("name").asText(),
+            component.has("version") ? component.get("version").asText() : null,
+            component.has("description") ? component.get("description").asText() : null,
+            component.has("type") ? component.get("type").asText() : null,
+            component.has("purl") ? component.get("purl").asText() : null,
+            component.has("bom-ref") ? component.get("bom-ref").asText() : null);
+
+        Mockito.when(generateSbomService.generate(Mockito.anyString()))
+            .thenReturn(mockSbom);
+    }
+
+    @Test
+    void testUpload_SpdxWithUnsupportedComponent_RecordsInSubmissionFailures() {
+        File sbomFile = new File(SPDX_WITH_UNSUPPORTED);
+
+        String productId = RestAssured.given()
+            .contentType(ContentType.MULTIPART)
+            .multiPart("cveId", TEST_VULN_ID)
+            .multiPart("file", sbomFile)
+            .when()
+            .post("/api/v1/products/upload-spdx")
+            .then()
+            .statusCode(202)
+            .contentType(ContentType.JSON)
+            .body("productId", notNullValue())
+            .extract()
+            .path("productId");
+
+        Assertions.assertNotNull(productId, "Product ID should not be null");
+        RestApiTestFixture.awaitSpdxProductProcessingComplete(productId);
+
+        RestAssured.given()
+            .when()
+            .get("/api/v1/products/" + productId)
+            .then()
+            .statusCode(200)
+            .contentType(ContentType.JSON)
+            .body("data.submittedCount", equalTo(2))
+            .body("data.submissionFailures", notNullValue())
+            .body("data.submissionFailures", hasSize(1))
+            .body("data.submissionFailures[0].name", equalTo("maven-lib"))
+            .body("data.submissionFailures[0].version", equalTo("2.0"))
+            .body("data.submissionFailures[0].error", containsString("Expects a container image purl with format pkg:oci/name@sha256:hash?repository_url=...&tag=..."));
+    }
+}
